+++ CHANGELOG.md
@@ -7,13 +7,19 @@ All notable changes to this project will be documented in this file.
 - Updated `IMAGE_OPTIONAL_HEADER` to support 64-bit and added missing `DllCharacteristics` Flags. (@ddash-ct)
 - Updated `IMAGE_FILE_HEADER.SizeOfOptionalHeader` to enable leveraging `sizeof()`. (@ddash-ct)
 - Changed log messages for file identification and misidentification to update phrasing for parsing groups vs parsing components. (@ddash-ct)
+- Added support for importing external parser components/groups within a parser configuration. (See [documentation](docs/ParserInstallation.md#grouping-parsers))
+- Added support for providing run configuration options to `FileObject.run_kordesii_decoder()` which will be passed 
+    along to `kordesii.run_ida()` when calling IDA. (This allows you to provide the new `is_64bit` option if necessary.)
 
 ### Fixed
 - Fixed glob pattern in Techanarchy wrapper. (@cccs-aa)
 - Fixed misspelling of "Characteristics" in `IMAGE_IMPORT_DESCRIPTOR`. (@ddash-ct)
-- Fixed infinte loop that can be caused due to a sub-parser throwing an `UnableToParse` exception. (@ddash-ct)
+- Fixed infinite loop that can be caused due to a sub-parser throwing an `UnableToParse` exception. (@ddash-ct)
 - Fixed bug in construct.Base64 adapter for build with unicode encoding types. (@ddash-ct)
-    
+- General fixes to improve support when running under Linux.
+    - Changed log configuration usage of `%LOCALAPPDATA%` for the log directory reported by `appdirs`.
+
+
 ## [3.1.0] - 2020-06-05
 
 ### Added